deleteTask() validation and status include in responses

+++ todo/src/main/java/com/todo/service/TaskService.java
@@ -52,6 +52,9 @@ public class TaskService {
     }
 
     public void deleteTask(Long id) {
+        if (!taskRepository.existsById(id)) {
+            throw new NotFoundException("Task not found with id: " + id);
+        }
         taskRepository.deleteById(id);
     }
